moving all logic files to logic folder
