+++ src/logic/Board/location.test.ts
